# openapi.yaml
openapi: 3.1.0
info:
  title: 포토스팟 추천 API
  description: 퍼스널컬러·체형 기반 사진 명소 추천. 소셜 로그인(/v1/auth/*)으로 받은 액세스 토큰을 Bearer로 보낸다.
  version: 1.2.0
paths:
...
  /v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh
      description: 갱신 토큰으로 새 토큰 쌍. 갱신 토큰은 한 번 쓰면 새것으로 바뀐다.
      operationId: refresh_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      description: 이 기기의 갱신 토큰을 무효화. 액세스 토큰은 수명(기본 1시간)이 끝나면 자동으로 만료
      operationId: logout_v1_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshIn'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/{provider}:
    post:
      tags:
      - auth
      summary: Login
      description: 서비스 토큰을 검증하고 우리 토큰을 발급. 처음이면 계정을 만든다.
      operationId: login_v1_auth__provider__post
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          enum:
          - google
          - apple
          - naver
          - kakao
          type: string
          title: Provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/me/identities/{provider}:
    post:
      tags:
      - auth
      - me
      summary: Link Identity
      description: 로그인된 계정에 다른 로그인 수단을 추가
      operationId: link_identity_v1_me_identities__provider__post
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          enum:
          - google
          - apple
          - naver
          - kakao
          type: string
          title: Provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginIn'
      responses:
        '204':
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - auth
      - me
      summary: Unlink Identity
      operationId: unlink_identity_v1_me_identities__provider__delete
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          enum:
          - google
          - apple
          - naver
          - kakao
          type: string
          title: Provider
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DeleteAccountIn'
              - type: 'null'
              title: Body
      responses:
        '204':
          description: Successful Response
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AppleDeletionProof:
      properties:
        identity_token:
          type: string
          maxLength: 16000
          minLength: 1
          title: Identity Token
        authorization_code:
          type: string
          maxLength: 4096
          minLength: 1
          title: Authorization Code
        nonce:
          type: string
          maxLength: 256
          minLength: 16
          title: Nonce
      type: object
      required:
      - identity_token
      - authorization_code
      - nonce
      title: AppleDeletionProof
...
    DeleteAccountIn:
      properties:
        apple:
          anyOf:
          - $ref: '#/components/schemas/AppleDeletionProof'
          - type: 'null'
      type: object
      title: DeleteAccountIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginIn:
      properties:
        token:
          type: string
          maxLength: 16000
          minLength: 1
          title: Token
          description: '구글·애플: ID 토큰(JWT). 네이버·카카오: 액세스 토큰'
        nonce:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Nonce
          description: '구글·애플: 요청 때 쓴 nonce (재생 공격 방지, 선택)'
        name:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Name
          description: '애플: 첫 로그인 때만 앱이 받는 이름'
        authorization_code:
          anyOf:
          - type: string
            maxLength: 4096
          - type: 'null'
          title: Authorization Code
          description: 'Apple 승인 코드: 서버에서 교환 후 암호화 보관'
      type: object
      required:
      - token
      title: LoginIn
...
    Problem:
      properties:
        type:
          type: string
          title: Type
          default: about:blank
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
      type: object
      required:
      - title
      - status
      title: Problem
      description: 오류 응답 (RFC 9457 형식).
...
    RefreshIn:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshIn
...
    TokenPair:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          const: bearer
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
          description: 액세스 토큰 수명(초)
        user_id:
          type: string
          title: User Id
        is_new:
          type: boolean
          title: Is New
          description: 이번 로그인으로 계정이 새로 만들어졌는지
        profile_exists:
          type: boolean
          title: Profile Exists
          description: false면 온보딩 화면으로
      type: object
      required:
      - access_token
      - refresh_token
      - expires_in
      - user_id
      - is_new
      - profile_exists
      title: TokenPair
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError